% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dem.R
\name{apply_threshold}
\alias{apply_threshold}
\title{Apply threshold to DEM}
\usage{
apply_threshold(dem, threshold = 923.667)
}
\arguments{
\item{dem}{A SpatRaster object with the DEM data.}

\item{threshold}{The threshold value (default is 923.667).}
}
\value{
A SpatRaster object with the threshold applied.
}
\description{
This function applies a threshold to the DEM data, setting values below the threshold to NA.
}
