% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{compute_watersheds}
\alias{compute_watersheds}
\title{Compute Watersheds from D8 Pointer and Pour Points}
\usage{
compute_watersheds(
  d8_pntr = system.file("extdata", "D8pointer.tif", package = "hydroModelling"),
  pour_pts = system.file("extdata", "snappedpp.shp", package = "hydroModelling"),
  output = tempfile(fileext = ".tif")
)
}
\arguments{
\item{d8_pntr}{A string specifying the file path to the D8 pointer raster.}

\item{pour_pts}{A string specifying the file path to the shapefile containing the snapped pour points.}

\item{output}{A string specifying the file path where the resulting watershed raster will be saved.}
}
\value{
The file path to the raster containing the computed watersheds.
}
\description{
This function calculates watershed areas based on the D8 pointer raster and pour points shapefile.
}
\examples{
watershed_path <- compute_watersheds(
  d8_pntr = "path/to/D8pointer.tif",
  pour_pts = "path/to/snappedpp.shp",
  output = "path/to/rw_watersheds.tif"
)

}
