% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_acc.R
\name{extract_streams}
\alias{extract_streams}
\title{Extract Streams from Flow Accumulation Raster}
\usage{
extract_streams(
  flow_accum_path = system.file("extdata", "D8FA.tif", package = "hydroModelling"),
  output_path = tempfile(fileext = ".tif"),
  threshold = 6000
)
}
\arguments{
\item{flow_accum_path}{A string specifying the file path to the flow accumulation raster. Default is a file located in the extdata directory of the package.}

\item{output_path}{A string specifying the file path where the extracted streams raster will be saved. Default is a temporary file.}

\item{threshold}{A numeric value representing the threshold for stream extraction. Cells in the flow accumulation raster with values above this threshold are considered as streams. Default is 6000.}
}
\value{
A string specifying the file path to the extracted streams raster.
}
\description{
This function extracts streams from a flow accumulation raster using a specified threshold. The streams are extracted using the whitebox package.
}
