% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshade.R
\name{fill_hillshade}
\alias{fill_hillshade}
\title{Fill Depressions in Hillshade DEM}
\usage{
fill_hillshade(
  hillshade_path = system.file("extdata", "hillshade.tif", package = "hydroModelling"),
  dist = 5,
  fill = TRUE
)
}
\arguments{
\item{hillshade_path}{The path to the input hillshade DEM file. Default is a file located in the extdata directory of the package.}

\item{dist}{The maximum search distance for breaching depressions. Default is 5.}

\item{fill}{Logical parameter to determine if the function should fill depressions. Default is TRUE.}
}
\value{
The path to the filled DEM file.
}
\description{
This function processes a hillshade DEM by breaching depressions and then filling them using the Wang and Liu method. The input hillshade DEM is taken from the extdata directory of the package. The function outputs a filled DEM that has had depressions breached and filled.
}
\examples{
\dontrun{
  filled_dem <- fill_hillshade(
    hillshade_path = system.file("extdata", "hillshade.tif", package = "hydroModelling"),
    dist = 10,
    fill = TRUE
  )
  print(filled_dem)
}



}
