% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_acc.R
\name{flow_acc_and_pointer}
\alias{flow_acc_and_pointer}
\title{Compute D8 Flow Accumulation and D8 Pointers}
\usage{
flow_acc_and_pointer(
  dem_path = system.file("extdata", "breached_hillshade.tif", package = "hydroModelling"),
  filled_hillshade_path = system.file("extdata", "filled_hillshade.tif", package =
    "hydroModelling"),
  flow_acc_output_path = tempfile(fileext = ".tif"),
  pointer_output_path = tempfile(fileext = ".tif")
)
}
\arguments{
\item{dem_path}{A string specifying the file path to the input DEM raster. This DEM is used for calculating the D8 pointers.}

\item{filled_hillshade_path}{A string specifying the file path to the input filled DEM raster, used for calculating the D8 flow accumulation.}

\item{flow_acc_output_path}{A string specifying the file path where the resulting D8 flow accumulation raster will be saved.}

\item{pointer_output_path}{A string specifying the file path where the resulting D8 pointer raster will be saved.}
}
\value{
A list with two elements:
\item{flow_acc_output}{A string representing the file path to the D8 flow accumulation raster.}
\item{pointer_output}{A string representing the file path to the D8 pointer raster.}
}
\description{
This function calculates the D8 flow accumulation and D8 pointers for a given digital elevation model (DEM) and filled DEM.
}
\details{
The D8 flow accumulation represents the total amount of flow that accumulates to each cell in the raster grid based on the D8 flow model,
which directs flow to one of the eight neighboring cells. The D8 pointer raster indicates the direction of flow from each cell.
}
\examples{
result <- flow_acc_and_pointer(
  dem_path = "path/to/dem.tif",
  filled_hillshade_path = "path/to/filled_dem.tif",
  flow_acc_output_path = "path/to/flow_accumulation.tif",
  pointer_output_path = "path/to/d8_pointer.tif"
)

}
