% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshade.R
\name{generate_hillshade}
\alias{generate_hillshade}
\title{Generate Hillshade from DEM}
\usage{
generate_hillshade(
  dem,
  output_path = tempfile(fileext = ".tif"),
  azimuth = 115
)
}
\arguments{
\item{dem}{A SpatRaster object representing the Digital Elevation Model (DEM).}

\item{output_path}{A string specifying the file path where the generated hillshade raster will be saved. Default is a temporary file path.}

\item{azimuth}{A numeric value representing the azimuth angle (in degrees) for the sun's position, used in creating the hillshade effect. Default is 115 degrees.}
}
\value{
A string specifying the file path to the generated hillshade raster.
}
\description{
This function generates a hillshade raster from a Digital Elevation Model (DEM). It first applies a threshold to the DEM, then generates the hillshade raster using the whitebox package.
}
\examples{
\dontrun{
dem <- terra::rast("path/to/dem.tif")
generate_hillshade(dem, output_path = "output/hillshade.tif", azimuth = 115)
}

}
