% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dem.R
\name{load_and_set_crs}
\alias{load_and_set_crs}
\title{Load DEM data and set CRS}
\usage{
load_and_set_crs(
  dem_path = system.file("extdata", "demRw.tif", package = "hydroModelling"),
  crs_code = "EPSG:4326"
)
}
\arguments{
\item{dem_path}{The path to the DEM file.}

\item{crs_code}{The CRS code to set for the DEM (default is "EPSG:4326").}
}
\value{
A SpatRaster object with the DEM data.
}
\description{
Load DEM data and set CRS
}
