% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dem.R
\name{plot_dem}
\alias{plot_dem}
\title{Plot DEM}
\usage{
plot_dem(dem = apply_threshold())
}
\arguments{
\item{dem}{A SpatRaster object with the DEM data.}
}
\description{
This function plots the DEM data using tmap.
}
