% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshade.R
\name{plot_hillshade}
\alias{plot_hillshade}
\title{Plot Hillshade Raster}
\usage{
plot_hillshade(
  dem,
  hillshade_path = system.file("extdata", "hillshade.tif", package = "hydroModelling")
)
}
\arguments{
\item{dem}{A SpatRaster object representing the Digital Elevation Model (DEM). This DEM will be used to generate the hillshade.}

\item{hillshade_path}{A string specifying the file path where the generated hillshade raster will be saved. Default is a file located in the extdata directory of the package.}
}
\value{
A tmap object representing the hillshade plot.
}
\description{
This function generates a hillshade raster from a Digital Elevation Model (DEM) and plots it using the tmap package. The hillshade raster is first created using the generate_hillshade function, and then plotted.
}
\examples{
\dontrun{
dem <- terra::rast("path/to/dem.tif")
plot_hillshade(dem, hillshade_path = "output/hillshade.tif")
}


}
