% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_acc.R
\name{plot_streams_and_pour_points}
\alias{plot_streams_and_pour_points}
\title{Plot Stream Networks and Pour Points}
\usage{
plot_streams_and_pour_points(
  streams_path = system.file("extdata", "raster_streams.tif", package = "hydroModelling"),
  snapped_pp_path = system.file("extdata", "snappedpp.shp", package = "hydroModelling")
)
}
\arguments{
\item{streams_path}{A string specifying the file path to the raster file containing the stream network. This raster will be visualized with a blue color palette.}

\item{snapped_pp_path}{A string specifying the file path to the shapefile containing the pour points. These points will be overlaid on the stream network in red.}
}
\value{
A tmap object that displays the stream network and pour points. This object can be printed or further customized.
}
\description{
This function creates a map visualization of stream networks and pour points using tmap.
It reads in a raster of stream networks and a shapefile of pour points, then plots them on the same map.
}
