% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{plot_watersheds}
\alias{plot_watersheds}
\title{Plot Watersheds, Hillshade, and Pour Points}
\usage{
plot_watersheds(
  hillshade_path = system.file("extdata", "hillshade.tif", package = "hydroModelling"),
  watershed_path = system.file("extdata", "rw_watersheds.tif", package =
    "hydroModelling"),
  pour_points_path = system.file("extdata", "pourpoints.shp", package = "hydroModelling")
)
}
\arguments{
\item{hillshade_path}{A string specifying the file path to the hillshade raster.}

\item{watershed_path}{A string specifying the file path to the watershed raster.}

\item{pour_points_path}{A string specifying the file path to the shapefile containing the pour points.}
}
\value{
A tmap object that displays the hillshade, watershed, and pour points. This object can be printed or further customized.
}
\description{
This function visualizes the watershed raster, hillshade raster, and pour points using tmap.
}
