% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_acc.R
\name{snap_pour_points}
\alias{snap_pour_points}
\title{Snap Pour Points to Stream Raster}
\usage{
snap_pour_points(
  pour_pts_path = system.file("extdata", "pourpoints.shp", package = "hydroModelling"),
  streams_path = system.file("extdata", "raster_streams.tif", package = "hydroModelling"),
  output_path = tempfile(fileext = ".tif"),
  snap_dist = 5e-04
)
}
\arguments{
\item{pour_pts_path}{A character string specifying the path to the pour points shapefile. Default is
the example shapefile provided by the \pkg{hydroModelling} package.}

\item{streams_path}{A character string specifying the path to the raster file containing stream
information. Default is the example raster file provided by the \pkg{hydroModelling} package.}

\item{output_path}{A character string specifying the path where the snapped pour points raster file
will be saved. Default is a temporary file with the \code{.tif} extension.}

\item{snap_dist}{A numeric value specifying the snapping distance. Pour points will be moved to the
nearest stream within this distance. Default is 0.0005.}
}
\value{
A character string with the path to the snapped pour points raster file.
}
\description{
This function snaps pour points to the nearest stream in a raster format. It uses the WhiteboxTools
to perform the snapping operation. The input pour points are adjusted to the nearest stream within
a specified snapping distance.
}
\examples{
# Example usage
snapped_pour_points <- snap_pour_points(
  pour_pts_path = system.file("extdata", "pourpoints.shp", package = "hydroModelling"),
  streams_path = system.file("extdata", "raster_streams.tif", package = "hydroModelling"),
  snap_dist = 0.001
)
print(snapped_pour_points)
}
