% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dem.R
\name{write_raster_with_crs}
\alias{write_raster_with_crs}
\title{Write raster with CRS}
\usage{
write_raster_with_crs(
  dem = load_and_set_crs(),
  output_path = "D:/Munster/Data Science w R/finalProject/demRw_30m_crs.tif"
)
}
\arguments{
\item{dem}{A SpatRaster object with the DEM data.}

\item{output_path}{The path to save the raster file.}
}
\description{
This function writes the raster data with CRS to a specified path.
}
